move args into VM
