--- app/src/main/java/com/sebastianvm/musicplayer/ui/artist/ArtistNavDestination.kt
@@ -1,4 +1,0 @@
-package com.sebastianvm.musicplayer.ui.artist
-
-data class ArtistArguments(val artistId: Long)
-
