+++ app/src/main/java/com/sebastianvm/musicplayer/ui/bottomsheets/context/PlaylistContextMenuViewModel.kt
@@ -103,6 +103,8 @@ class PlaylistContextMenuViewModel @Inject constructor(
     }
 }
 
+data class PlaylistContextMenuArguments(val playlistId: Long)
+
 
 data class PlaylistContextMenuState(
     override val listItems: List<ContextMenuItem>,
