--- app/src/main/java/com/sebastianvm/musicplayer/ui/bottomsheets/mediaartists/ArtistsBottomSheetNavDestination.kt
@@ -1,5 +1,0 @@
-package com.sebastianvm.musicplayer.ui.bottomsheets.mediaartists
-
-import com.sebastianvm.musicplayer.model.MediaWithArtists
-
-data class ArtistsMenuArguments(val mediaType: MediaWithArtists, val mediaId: Long)
