--- app/src/main/java/com/sebastianvm/musicplayer/ui/bottomsheets/sort/SortBottomSheetNavDestination.kt
@@ -1,3 +1,0 @@
-package com.sebastianvm.musicplayer.ui.bottomsheets.sort
-
-data class SortMenuArguments(val listType: SortableListType)
