--- app/src/main/java/com/sebastianvm/musicplayer/ui/library/tracklist/TrackListNavDestination.kt
@@ -1,6 +1,0 @@
-package com.sebastianvm.musicplayer.ui.library.tracklist
-
-import com.sebastianvm.musicplayer.player.TrackList
-
-
-data class TrackListArguments(val trackList: TrackList?)
