--- app/src/main/java/com/sebastianvm/musicplayer/ui/playlist/TrackSearchNavDestination.kt
@@ -1,4 +1,0 @@
-package com.sebastianvm.musicplayer.ui.playlist
-
-data class TrackSearchArguments(val playlistId: Long)
-
